mpi: simplify async path (#396)
In this PR I simplify the mechanism to handle MPI async messages. Most
importantly, there is a (relatively) fast path for when there are no
async messages at all.

In the process, I refine the generation of isend/irecv request ids so
that we do not need to track them in a map.

Lastly, with this improvements we can get rid of the MpiMessageBuffer
that was adding more complexity than needed.