mpi: make all-to-all implementation locality-aware (#398)
This PR implements a locality-aware AllToAll collective primitive.
Similarly to what we have done in the past for other colelctives like
broadcast and reduce, the current implementation aims to minimise the
number of cross-VM messages.

Differently to broadcast or reduce, this implementation adds _redundant_
local messages, to avoid remove messages. In general, an allToAll can be
thought of as each rank sending to all ranks, and receiving from all
ranks.

Now, we split it in two parts:
1. Co-located ranks send/recv in the same fashion as before.
2. Ranks send all their non-local messages to the local leader, wihch
   then batches them into _one_ cross-VM message per remote host, and
   then receives the same message from each other host, and sends the
   messages downstream.

In summary, we add two extra local messages per message, so now we send
O(2*n^2) messages, but we reduce the number of cross-VM messages from
O(2 * num_crossvm_links) to O(2 * num_hosts).

Our implementation is optimal in terms of cross-vm messages.