mpi: temporarily revert locality-aware all-to-all (#402)
It seems that the optimisation introduced in #398 is not such a
low-hanging fruit. Preliminary evidence suggests that replacing one
cross-VM message by two additional local messages + contention on the
local leaders does not quite add up.

For the time being, I will revert the optimisation, but leave the
commented code, until we have more decisive evidence.