planner(mpi): preload MPI SCALE_CHANGE decisions
When scheduling a NEW app with the `ismpi` flag set, we know that this
request will be quickly followed-up by a SCALE_CHANGE request with
`worldSize` - 1 messages. As a consequence, and to guarantee optimal
bin-packing on-par with Slurm, we can schedule `worldSize` messages when
the first request arrives, dispatch just the first one, and preload the
`worldSize` - 1 remaining.