transport(tcp): remove sendMany function
Turns out that sendMany becomes very convoluted when we need to handle
partial writes. As it is not very clear the benefit that we are getting
from it, and we may get away without it, we remove it in this PR.

I also fix the handling of partial writes for the sendOne method.