transport(tcp): make send sockets blocking
Under high-load we are seeing some ::send operations fail with
EAGAIN/EWOULDBLOCK. This is because we set send sockets as non-blocking,
but we do not handle the EAGAIN/EWOULDBLOCK situation in the send loop.
As a first iteration, we attempt setting the send sockets as blocking.

Note that, irrespective of wether send sockets block or not, partial
writes may still happen.