planner: simplify accounting during migration (#427)
There was a bug in the accounting logic during a DIST_CHANGE request. At
this point, we need to release the slots of the to-be migrated function,
and claim them in the migrated-to hosts.

The current logic assumed that, given two decisions, they would always
start being similar, and then diverge (or even diverge all the way
through).

This is not the case. Take for example the following situation where A
has 5 slots:

```
Old, New
B, A
B, A
B, A
C, A
C, A
A, C
A, C
```

Using our accounting logic, for a while we would have 7 slots assigned
to A. This would break our assertions (did not catch it because it only
happens in Release builds), but was breaking the MPI port assignation.