migration: fix re-entrant bug
This PR fixes a rare race condition that only happened in enviornments
where the same app is migrated many times.

In particular, this bug only appeared when the same application migrated
away from one host, and then migrated back into it. Migrating into a new
host (wrt the previous scheduling decision) requires one of the
migrated-to ranks to run the world initialisation to set the
local-remote leaders and in-memory queues. However, the second migration
above was not triggering the "new world" migration procedure because the
world had lingered in the per-node registry.

This bug materialised in applications having an old version of the
host-port mappings, and failing to start.

The fix involves knowing when we are evicting a host for a given world
id, and clearing it from the registry if so.