mpi: fix setting local ranks
In very constrained environments like GHA it may sometimes happen that
some ranks _completely_ finish (i.e. call destroy) before others have
even called Init.

With our current implementation, this meant that the world may have been
removed from the registry, even though there were still active local
ranks to run.

To fix this, we set the number of active local ranks once, at the
beginning, and decrement it on a per-thread basis. Note that this does
not invaliate the fix to the previous race condition because, in fact,
what we fixed is that we now _always_ init a world when we execute in
it.