scheduler: tag executors with app id (#434)
Thus far, we tagged executors with the (user, function) pair. This is
needed in multi-threaded environments, where different threads must
share the same executor. However, using only the (user, function) pair
prevents the execution concurrent applications with the same pair.

Instead, the correct thing to do is to track applications using the
(user, function) pair in conjunction with the application id.